Test: Split SqlServerDatabaseCreator tests
